+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Next
+
+- **[Feature]** Define `DefineButtonSound` tag.
+
 # 0.8.0 (2019-07-08)
 
 - **[Breaking change]** Require `language` to be defined in `DefineFontInfo`.